requirements.txt
aniso8601==3.0.2
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
coverage==4.5.1
Flask==1.0.2
Flask-JWT-Extended==3.12.1
flask-restplus==0.11.0
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
more-itertools==4.3.0
nose==1.3.7
pluggy==0.7.1
py==1.6.0
PyJWT==1.6.4
pytest==3.7.4
pytest-cov==2.5.1
python-coveralls==2.9.1
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
